Sequential file to upload as picture
file size is 640x480 with RGB data incrementing by 1 from 0x0000 to 0x03ff the starts back at 0x0000. This is for troubleshooting purposes during design. 
When a line is read form the SDRAM if end of page is reached SDRAM controller needs to switch to the next page on the SDRAM. Using this file I was able to ensure the switch was happening as intended.